pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smrlib"
version = "0.1.0"
description = "Personal utility library for Python projects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.6",
    "gspread>=6.0.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "python-dotenv>=1.0.0",
    "httplib2>=0.22.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-clarity>=1.0.1",
    "black>=25.9.0",
    "ruff>=0.14.1",
    "pre-commit>=4.0.0",
    "pyright>=1.1.407",
]

[tool.hatch.build.targets.wheel]
packages = ["smrlib"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
# Abort on first failure/error by default
# -x: stop on first failure
# -v: verbose (show test names)
# -s: show print/logger output (disable capture)
# --tb=short: short traceback format
addopts = "-x -v -s --tb=short"
# Log configuration: show all logger output during tests
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)s %(message)s"
log_cli_date_format = "%H:%M:%S"

[tool.pyright]
include = ["smrlib", "tests"]
exclude = ["**/__pycache__", ".venv", "build", "dist"]
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.10"
pythonPlatform = "All"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["smrlib"]
extend-exclude = [
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "A",   # flake8-builtins
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "PL",  # pylint
]
ignore = [
    "E501",    # Line too long (handled by black)
    "PLE1205", # Too many arguments for logging format string
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments
    "PLR0915", # Too many statements
    "PLR5501", # Consider using elif
    "PLR2004", # Magic value used in comparison
    "PLW0603", # Using the global statement
    "S101",    # Use of assert detected (pytest uses assertions)
    "ARG002",  # Unused method argument
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["smrlib"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

